add RVExtensionVersion to extensions
extract NativeMethods to own class
fix posible problem with that Charset is different in external functions
only show debugger when dll is build with debug enabled
imporve Build Tool
fix some genereal build warnings
Prepare for appveyor.
