+++ extensions/CLib/CLib.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26403.7
+VisualStudioVersion = 15.0.26430.16
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CLib", "CLib\CLib.csproj", "{3B6CA58C-9DB3-4256-8A10-BC7101D1EE8A}"
 EndProject
@@ -13,22 +13,40 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CLibWebSocket", "CLibWebSoc
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3B6CA58C-9DB3-4256-8A10-BC7101D1EE8A}.Debug|x64.ActiveCfg = Debug|x64
+		{3B6CA58C-9DB3-4256-8A10-BC7101D1EE8A}.Debug|x64.Build.0 = Debug|x64
+		{3B6CA58C-9DB3-4256-8A10-BC7101D1EE8A}.Debug|x86.ActiveCfg = Debug|x86
+		{3B6CA58C-9DB3-4256-8A10-BC7101D1EE8A}.Debug|x86.Build.0 = Debug|x86
 		{3B6CA58C-9DB3-4256-8A10-BC7101D1EE8A}.Release|x64.ActiveCfg = Release|x64
 		{3B6CA58C-9DB3-4256-8A10-BC7101D1EE8A}.Release|x64.Build.0 = Release|x64
 		{3B6CA58C-9DB3-4256-8A10-BC7101D1EE8A}.Release|x86.ActiveCfg = Release|x86
 		{3B6CA58C-9DB3-4256-8A10-BC7101D1EE8A}.Release|x86.Build.0 = Release|x86
+		{C780B540-D5AA-4C4E-A89B-E50D5DBF76A4}.Debug|x64.ActiveCfg = Debug|x64
+		{C780B540-D5AA-4C4E-A89B-E50D5DBF76A4}.Debug|x64.Build.0 = Debug|x64
+		{C780B540-D5AA-4C4E-A89B-E50D5DBF76A4}.Debug|x86.ActiveCfg = Debug|x86
+		{C780B540-D5AA-4C4E-A89B-E50D5DBF76A4}.Debug|x86.Build.0 = Debug|x86
 		{C780B540-D5AA-4C4E-A89B-E50D5DBF76A4}.Release|x64.ActiveCfg = Release|x64
 		{C780B540-D5AA-4C4E-A89B-E50D5DBF76A4}.Release|x64.Build.0 = Release|x64
 		{C780B540-D5AA-4C4E-A89B-E50D5DBF76A4}.Release|x86.ActiveCfg = Release|x86
 		{C780B540-D5AA-4C4E-A89B-E50D5DBF76A4}.Release|x86.Build.0 = Release|x86
+		{6EC8493C-6803-487C-8061-0F8C4D33A7E6}.Debug|x64.ActiveCfg = Debug|x64
+		{6EC8493C-6803-487C-8061-0F8C4D33A7E6}.Debug|x64.Build.0 = Debug|x64
+		{6EC8493C-6803-487C-8061-0F8C4D33A7E6}.Debug|x86.ActiveCfg = Debug|x86
+		{6EC8493C-6803-487C-8061-0F8C4D33A7E6}.Debug|x86.Build.0 = Debug|x86
 		{6EC8493C-6803-487C-8061-0F8C4D33A7E6}.Release|x64.ActiveCfg = Release|x64
 		{6EC8493C-6803-487C-8061-0F8C4D33A7E6}.Release|x64.Build.0 = Release|x64
 		{6EC8493C-6803-487C-8061-0F8C4D33A7E6}.Release|x86.ActiveCfg = Release|x86
 		{6EC8493C-6803-487C-8061-0F8C4D33A7E6}.Release|x86.Build.0 = Release|x86
+		{9573C0AE-7B68-4B63-B7DD-E982E62F6086}.Debug|x64.ActiveCfg = Debug|x64
+		{9573C0AE-7B68-4B63-B7DD-E982E62F6086}.Debug|x64.Build.0 = Debug|x64
+		{9573C0AE-7B68-4B63-B7DD-E982E62F6086}.Debug|x86.ActiveCfg = Debug|x86
+		{9573C0AE-7B68-4B63-B7DD-E982E62F6086}.Debug|x86.Build.0 = Debug|x86
 		{9573C0AE-7B68-4B63-B7DD-E982E62F6086}.Release|x64.ActiveCfg = Release|x64
 		{9573C0AE-7B68-4B63-B7DD-E982E62F6086}.Release|x64.Build.0 = Release|x64
 		{9573C0AE-7B68-4B63-B7DD-E982E62F6086}.Release|x86.ActiveCfg = Release|x86
