+++ extensions/CLib/CLibCompression/DllEntry.cs
@@ -4,6 +4,7 @@ using System.Runtime.InteropServices;
 using System.Text;
 using System.Reflection;
 using System.Diagnostics;
+using CLib;
 
 namespace CLibCompression
 {
@@ -12,11 +13,31 @@ namespace CLibCompression
         private const int WindowSize = 1 << 11;
         private const int MinMatchLength = 2;
         private const uint MaxMatchLength = (1 << 4) - MinMatchLength;
-
+#if WIN64
+		[DllExport("RVExtensionVersion")]
+#else
+		[DllExport("_RVExtensionVersion@6", CallingConvention.StdCall)]
+#endif
+		public static void RVExtensionVersion(StringBuilder output, int outputSize)
+		{
+			outputSize--;
+			var executingAssembly = Assembly.GetExecutingAssembly();
+			try
+			{
+				string location = executingAssembly.Location;
+				if (location == null)
+					throw new Exception("Assembly location not found");
+				output.Append(FileVersionInfo.GetVersionInfo(location).FileVersion);
+			}
+			catch (Exception e)
+			{
+				output.Append(e.Message);
+			}
+		}
 #if WIN64
         [DllExport("RVExtension")]
 #else
-        [DllExport("_RVExtension@12", CallingConvention.StdCall)]
+		[DllExport("_RVExtension@12", CallingConvention.StdCall)]
 #endif
         public static void RVExtension(StringBuilder output, int outputSize, [MarshalAs(UnmanagedType.LPStr)] string input)
         {
