+++ extensions/CLib/CLibLogging/DllEntry.cs
@@ -2,6 +2,8 @@
 using System.Runtime.InteropServices;
 using System.Text;
 using System.IO;
+using System.Reflection;
+using System.Diagnostics;
 
 namespace CLibLogging
 {
@@ -12,11 +14,32 @@ namespace CLibLogging
         {
             startTime = DateTime.Now.ToString("yyyy-MM-dd_HH-mm-ss");
         }
+#if WIN64
+		[DllExport("RVExtensionVersion")]
+#else
+		[DllExport("_RVExtensionVersion@6", CallingConvention.StdCall)]
+#endif
+		public static void RVExtensionVersion(StringBuilder output, int outputSize)
+		{
+			outputSize--;
+			var executingAssembly = Assembly.GetExecutingAssembly();
+			try
+			{
+				string location = executingAssembly.Location;
+				if (location == null)
+					throw new Exception("Assembly location not found");
+				output.Append(FileVersionInfo.GetVersionInfo(location).FileVersion);
+			}
+			catch (Exception e)
+			{
+				output.Append(e.Message);
+			}
+		}
 
 #if WIN64
         [DllExport("RVExtension")]
 #else
-        [DllExport("_RVExtension@12", CallingConvention.StdCall)]
+		[DllExport("_RVExtension@12", CallingConvention.StdCall)]
 #endif
         public static void RVExtension(StringBuilder output, int outputSize, [MarshalAs(UnmanagedType.LPStr)] string input)
         {
