+++ extensions/CLib/CLibWebSocket/DllEntry.cs
@@ -7,9 +7,31 @@ using System.Text;
 namespace CLibWebSocket
 {
     public class DllEntry
-    {
+	{
 #if WIN64
-        [DllExport("RVExtension")]
+		[DllExport("RVExtensionVersion")]
+#else
+		[DllExport("_RVExtensionVersion@6", CallingConvention.StdCall)]
+#endif
+		public static void RVExtensionVersion(StringBuilder output, int outputSize)
+		{
+			outputSize--;
+			var executingAssembly = Assembly.GetExecutingAssembly();
+			try
+			{
+				string location = executingAssembly.Location;
+				if (location == null)
+					throw new Exception("Assembly location not found");
+				output.Append(FileVersionInfo.GetVersionInfo(location).FileVersion);
+			}
+			catch (Exception e)
+			{
+				output.Append(e.Message);
+			}
+		}
+
+#if WIN64
+		[DllExport("RVExtension")]
 #else
         [DllExport("_RVExtension@12", CallingConvention.StdCall)]
 #endif
